Install the Plotly JupyterLab extension alongside Plotly.py
Fixes #84

This manually installs the Plotly JupyterLab extension as per the
instructions here:

https://conda-forge.org/docs/maintainer/knowledge_base.html#jupyterlab-extensions

If the user has JupyterLab installed in the same conda environment as
plotly, the extension will be built by JupyterLab the next time it is
opened and be available for use. Node and JupyterLab are listed in the
run_constrained section so that these packages are not hard dependencies
of the Plotly conda package.